Change the backend and frontend interaction
The backend doesn't implement WebXdc anymore, but WebXdcMulti,
which sends multiple messsages to the frontend at once in
case of a (re)connection.

We also introduce some tests for the webxdc frontend. To this end
we've abstracted out a transport so we can put in a fake transport
during tests.